// src/api/triton/types.gen.ts
export type DocumentListResponse = {
    /**
     * Documents
     */
    documents: Array<DocumentResponse>;
    /**
     * Total
     */
    total: number;
    /**
     * Page
     */
    page: number;
    /**
     * Page Size
     */
    page_size: number;
};

/**
 * DocumentResponse
 */
export type DocumentResponse = {
    /**
     * Id
     */
    id: string;
    /**
     * Type
     */
    type: string;
    /**
     * Source Url
     */
    source_url: string | null;
    /**
     * Title
     */
    title: string | null;
    /**
     * Content
     */
    content: string;
    /**
     * Metadata
     */
    metadata?: {
        [key: string]: unknown;
    } | null;
    /**
     * Created At
     */
    created_at: string;
};